refactor: dnd context and data management
