+++ packages/extension/src/newtab/DndBanner.tsx
@@ -4,15 +4,13 @@ import XIcon from '@dailydotdev/shared/icons/x.svg';
 import DndContext from './DndContext';
 
 export default function DndBanner(): ReactElement {
-  const { setDndSettings } = useContext(DndContext);
+  const { onDndSettings } = useContext(DndContext);
 
-  const turnOff = () => setDndSettings(null);
+  const turnOff = () => onDndSettings(null);
 
   return (
-    <div className="flex relative flex-col laptop:flex-row laptop:justify-center items-start laptop:items-center laptop:p-0 py-3 pr-12 pl-3 laptop:h-8 typo-footnote bg-theme-bg-onion">
-      <div>
-        <strong>Do not disturb mode is active</strong>
-      </div>
+    <div className="flex relative laptop:fixed z-3 flex-col laptop:flex-row laptop:justify-center items-start laptop:items-center laptop:p-0 py-3 pr-12 pl-3 w-full laptop:h-8 typo-footnote bg-theme-bg-onion">
+      <strong>Do not disturb mode is active</strong>
       <Button
         buttonSize="xsmall"
         className="mt-2 laptop:mt-0 laptop:ml-4 btn-primary"
